Solidity code for a blockchain-based P2P betting system
A project showcasing three models of blockchain-based peer-to-peer (P2P) betting systems implemented as smart contracts in Solidity. The models simulate different betting scenarios using Ethereum testnet and Metamask wallet, and are tested both in the Remix environment and through Python scripts with the help of web3 and eth-tester libraries.